refactor(animations): Refactor node animation code
