+++ bagl-engine/src/main/java/com/adrienben/games/bagl/engine/animation/NodeKeyFrame.java
@@ -3,14 +3,21 @@ package com.adrienben.games.bagl.engine.animation;
 import org.joml.Quaternionf;
 import org.joml.Vector3f;
 
-public class KeyFrame {
+import java.util.Optional;
+
+/**
+ * Represent the state of a node at a given time.
+ *
+ * @author adrien
+ */
+public class NodeKeyFrame {
 
     private final float time;
     private Vector3f translation;
     private Quaternionf rotation;
     private Vector3f scale;
 
-    public KeyFrame(final float time) {
+    public NodeKeyFrame(final float time) {
         this.time = time;
     }
 
@@ -18,24 +25,24 @@ public class KeyFrame {
         return time;
     }
 
-    public Vector3f getTranslation() {
-        return translation;
+    public Optional<Vector3f> getTranslation() {
+        return Optional.ofNullable(translation);
     }
 
     public void setTranslation(Vector3f translation) {
         this.translation = translation;
     }
 
-    public Quaternionf getRotation() {
-        return rotation;
+    public Optional<Quaternionf> getRotation() {
+        return Optional.ofNullable(rotation);
     }
 
     public void setRotation(Quaternionf rotation) {
         this.rotation = rotation;
     }
 
-    public Vector3f getScale() {
-        return scale;
+    public Optional<Vector3f> getScale() {
+        return Optional.ofNullable(scale);
     }
 
     public void setScale(Vector3f scale) {
